requirements.txt
celluloid==0.2.0
GPy==1.10.0
#GPy==1.13.2
graphtools==1.5.3
ipdb==0.13.13
ipython==8.12.3
line_profiler==4.1.2
matplotlib==3.7.1
memory_profiler==0.61.0
ml_collections==0.1.1
numpy==1.23.5
pandas==2.0.3
phate==1.0.11
Pillow==9.4.0
#Pillow==11.0.0
pods==0.1.14
POT==0.9.3
prefetch_generator==1.0.3
pyro_ppl==1.8.6
PyYAML==6.0
#PyYAML==6.0.2
rich==13.9.2
#scanpy==1.10.3
scikit_learn==1.3.2
#scipy==1.14.1
#scprep==1.2.3
seaborn==0.13.2
setuptools==65.6.3
six==1.16.0
termcolor==2.4.0
torch==2.2.1
torch_ema==0.3
tqdm==4.64.1
#tqdm==4.65.0
typing_extensions==4.12.2
